Fixing build on macOS and sanitize the CMake so we can build with modern compilers
