+++ src/pydegensac/degensac/DegUtils.c
@@ -1,4 +1,4 @@
-#undef __STRICT_ANSI__
+//#undef __STRICT_ANSI__
 #include "DegUtils.h"
 #include "rtools.h"
 #include "Htools.h"
@@ -12,6 +12,7 @@
 #include <math.h>
 //#include <mex.h>
 #include <time.h>
+#include <matutls/ccmath.h>
 
 /*function [deg, H, inl] = checksample(F, u7, th)
