+++ src/pydegensac/degensac/Htools.c
@@ -2,12 +2,15 @@
 
 #include <memory.h>
 
+#include <stdio.h>
+
 #include "lapwrap.h"
 #include "../matutls/matutl.h"
 #include "utools.h"
 #include "rtools.h"
 
 #include "Htools.h"
+#include "math.h"
 
 #ifndef MAX
 #define MAX(i,j) ( (i)<(j) ? (j):(i) )
