+++ src/pydegensac/matutls/matconsts.c
@@ -1,0 +1,24 @@
+//
+// Created by Fan Jiang on 1/11/21.
+//
+
+/* Variables used in extended precision arithmetic */
+#include "ccmath.h"
+unsigned short m_sgn=0x8000,m_exp=0x7fff;
+short bias=16383;
+int itt_div=2,k_tanh=5;
+int ms_exp=21,ms_hyp=25,ms_trg=31;
+short max_p=16*XDIM,k_lin= -8*XDIM;
+short d_bias=15360,d_max=2047,d_lex=12;
+struct xpr zero= {{0x0,0x0}};
+struct xpr one= {{0x3fff,0x8000}};
+struct xpr two= {{0x4000,0x8000}};
+struct xpr ten= {{0x4002,0xa000}};
+struct xpr x_huge= {{0x7fff,0x0}};
+
+struct xpr pi4= {{0x3FFE,0xC90F,0xDAA2,0x2168,0xC234,0xC4C6,0x628B,0x80DC}};
+struct xpr pi2= {{0x3FFF,0xC90F,0xDAA2,0x2168,0xC234,0xC4C6,0x628B,0x80DC}};
+struct xpr pi= {{0x4000,0xC90F,0xDAA2,0x2168,0xC234,0xC4C6,0x628B,0x80DC}};
+struct xpr ee= {{0x4000,0xADF8,0x5458,0xA2BB,0x4A9A,0xAFDC,0x5620,0x273D}};
+struct xpr ln2= {{0x3FFE,0xB172,0x17F7,0xD1CF,0x79AB,0xC9E3,0xB398,0x3F3}};
+struct xpr srt2= {{0x3FFF,0xB504,0xF333,0xF9DE,0x6484,0x597D,0x89B3,0x754B}};
